Rename generator-base-entity to generator-base-entities.
