+++ generators/java/index.cjs
@@ -19,7 +19,7 @@
 /* eslint-disable consistent-return */
 const chalk = require('chalk');
 const { generateMixedChain } = require('../../lib/support/mixin.cjs');
-const GeneratorBaseEntity = require('../generator-base-entity.cjs');
+const GeneratorBaseEntity = require('../generator-base-entities.cjs');
 const {
   INITIALIZING_PRIORITY,
   PROMPTING_PRIORITY,
