+++ lib/index.mjs
@@ -3,4 +3,4 @@ export { default as utils } from '../generators/utils.js';
 
 export { default as GeneratorBase } from '../generators/generator-base.js';
 export { default as GeneratorBaseBlueprint } from '../generators/generator-base-blueprint.js';
-export { default as GeneratorBaseWithEntity } from '../generators/generator-base-entity.cjs';
+export { default as GeneratorBaseEntities } from '../generators/generator-base-entities.cjs';
